Added video playback tests
