--- requirements.txt
@@ -1,4 +1,0 @@
-requests
-termcolor
-dnspython
-pyopenssl
